refactor: (원시값 포장) 로또에 사용되는 번호를 Number로 포장
