+++ README.md
@@ -27,6 +27,6 @@
 - [x] 지난 주 당첨 번호를 입력한다.
 - [x] 당첨 통계를 출력한다.
 - [x] 상금 관련 필드를 enum에 넣는다.
-- [ ] 원시값을 포장한다.
+- [x] 원시값을 포장한다.
 - [ ] View 영역을 명확히 구분한다.
 - [ ] Generator에 정적 팩토리 메서드를 활용해본다.
