+++ src/main/java/lotto/domain/Lotto.java
@@ -2,13 +2,14 @@ package lotto.domain;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class Lotto {
-    private final List<Integer> numbers;
+    private final List<Number> numbers;
 
     public Lotto(List<Integer> numbers) {
-        this.numbers = numbers;
+        this.numbers = numbers.stream().map(Number::new).collect(Collectors.toList());
     }
 
     public boolean hasSize(int number) {
@@ -16,11 +17,11 @@ public class Lotto {
     }
 
     public boolean isValidRangeNumbers() {
-        return this.numbers.stream().noneMatch(number -> number < 1 || number > 45);
+        return this.numbers.stream().allMatch(Number::isValidRangeNumber);
     }
 
     public boolean isSorted() {
-        return IntStream.range(0, numbers.size() - 1).allMatch(i -> numbers.get(i) < numbers.get(i + 1));
+        return IntStream.range(0, numbers.size() - 1).allMatch(i -> numbers.get(i).isSmallerThan(numbers.get(i + 1)));
     }
 
     public String toString() {
