+++ src/main/java/lotto/domain/Number.java
@@ -1,0 +1,36 @@
+package lotto.domain;
+
+import java.util.Objects;
+
+public class Number {
+    private final int value;
+
+    public Number(int value) {
+        this.value = value;
+    }
+
+    public boolean isValidRangeNumber() {
+        return value >= 1 && value <= 45;
+    }
+
+    public boolean isSmallerThan(Number other) {
+        return value < other.value;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o == null || getClass() != o.getClass()) return false;
+        Number number = (Number) o;
+        return value == number.value;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(value);
+    }
+
+    @Override
+    public String toString() {
+        return String.valueOf(value);
+    }
+}
