+++ src/test/java/lotto/domain/ParserTest.java
@@ -1,6 +1,5 @@
-package lotto.util;
+package lotto.domain;
 
-import lotto.domain.Lotto;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
